backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
rembg
scikit-learn
mediapipe==0.10.9
onnxruntime
webcolors
requests
